latest

+++ forceback/background.js
@@ -2,51 +2,66 @@ let url_map = new Map();
 let redirect_map = new Map();
 
 chrome.tabs.onUpdated.addListener(async (tabId, changeInfo, tab) => {
-	console.log("onUpdated(" + tabId + ", " + tab.url);
-	
 	if(typeof tabId != "undefined"
+			&& typeof changeInfo != "undefined"
+		 	&& typeof changeInfo.status != "undefined"
+			&& changeInfo.status == "loading"
 			&& tab != "undefined"
 			&& tab.url != "undefined") {
-		console.log("onUpdated(" + tabId + ", " + tab.url);
-		
 		if(url_map.has(tabId)) {
 			let last_url = url_map.get(tabId);
 			
 			if(last_url != tab.url) {
-				console.log("New url: " + tab.url);
+				console.log("tabs.onUpdated(" + tabId + ") - " + tab.url);
 				
 				redirect_map.set(tabId, 0);
 				url_map.set(tabId, tab.url);			
 			}
+		} else {
+			console.log("tabs.onUpdated(" + tabId + ") - " + tab.url);
+			url_map.set(tabId, tab.url);
+			redirect_map.set(tabId, 0);
 		}
-	}
+	}	
 	
 	return true;
 });
 
 chrome.tabs.onRemoved.addListener(async (tabId, removeInfo) => {
 	if(typeof tabId != "undefined") {
-		url_map.delete(tabId);
-		redirect_map.delete(tabId);	
+		console.log("tabs.onRemoved(" + tabId + ")");
+		
+		if(url_map.has(tabId)) {
+			url_map.delete(tabId);
+		}
+		
+		if(redirect_map.has(tabId)) {
+			redirect_map.delete(tabId);
+		}			
 	}	
 	
 	return true;
 });
 
 chrome.history.onVisited.addListener(async (result) => {
-	let url = result.url;
-	let lastVisitTime = result.lastVisitTime;
+	if(typeof result != "undefined"
+			&& typeof result.url != "undefined") {
+		let url = result.url;
+		
+		chrome.tabs.getCurrent(async (tab) => {
+			if(typeof tab != "undefined"
+					&& typeof tab.id != "undefined") {
+				console.log("history.onVisited(" + tab.id + ") visited " + result.url);			
+			}
+		});
+	}
 	
-	console.log()
+	
+	
 	return true;
 });
 
 chrome.webNavigation.onHistoryStateUpdated.addListener(async (details) => {
-	var a = true;
-	if(a) {
-		return true;	
-	}	
-
 	if (typeof details !== "undefined"
 			&& typeof details.tabId !== "undefined"
 			&& typeof details.transitionQualifiers !== "undefined"
@@ -59,7 +74,7 @@ chrome.webNavigation.onHistoryStateUpdated.addListener(async (details) => {
 					redirect_map.set(details.tabId, 1);
 				}
 				
-				console.log("client_redirect, setting redirect_count to " + redirect_map.get(details.tabId) + " for tab " + details.tabId);
+				console.log("webNavigation.onHistoryStateUpdated(" + details.tabId + ") - redirect " + redirect_map.get(details.tabId));
 			}
 		}
 	}
